requirements.txt
germalemma==0.1.3
SoMaJo==2.0.1
SoMeWeTa==1.7.1
spacy==2.2.3
lxml==4.4.2
pandas==0.25.3
treetaggerwrapper==2.3
flair==0.4.4
syntok==1.2.2
python-igraph==0.7.1.post6
syntok==1.2.2
demorphy==1.0
torch==1.3.1+cpu
torchvision==0.4.2+cpu
